+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased
+## v7.1.0 - 2026-09-07
 
 ### Changed